+++ NEWS.md
@@ -2,8 +2,52 @@
 
 ## Changes
 
+* Methods for the *tinyplot* package were added.
+
+* `estimate_slopes()` now also gets the `estimate` argument, to specify how to
+  estimate over non-focal terms.
+
+* Improvements to `estimate_contrasts()`:
+
+  * `comparison = "inequality"` now also works when contrasting slopes for a
+    numeric predictor.
+
+  * New option `comparison = "inequality_ratio"` for `estimate_contrast()`,
+    which computes the ratio of marginal effects inequality measures across
+    groups.
+
+  * For contrasts, the message about the units of contrasts ("in %-points") has
+    been removed, because this was slightly misleading. The units were in
+    %-points if multiplied by 100, but this multiplication was not done in the
+    output.
+
+  * `estimate_contrasts()` for slopes now allows additional grouping of contrasts
+    using the `by` argument together with the `comparison` argument by specifying
+    the grouping variable in the formula, e.g. `contrast = c("x", "group")` and
+    `~ pairwise | group`.
+
 * `estimate_expectation()` and `estimate_relation()` now support objects of
   class `htest`.
+
+* `estimate_grouplevel()` now supports models from package *coxme*.
+
+* Documentation of the `display()` method for *modelbased* objects has been
+  added.
+
+* Improved documentation and improved informative messages.
+
+* Message about unreliable standard errors (for certain models, when predicting
+  random effects) was removed for now, as it is uncertain whether the standard
+  errors were unreliable.
+
+## Bug fixes
+
+* Fixed issue with `by` in `estimate_contrasts()` when `comparison` was
+  `"inequality"`.
+
+* Some comparison options, like `"helmert"` or `"poly"`, could not be specified
+  as string-value, only as formula. This has been fixed, so they can now be
+  specified as string-value, too.
 
 # modelbased 0.12.0